+++ packages/core/src/api/updateSession.ts
@@ -12,12 +12,14 @@ export const updateSession = async <DefaultUser extends User = User>({
     session: sessionInit,
     redirectTo: redirectToInit,
     skipCSRFCheck = false,
+    doubleSubmitToken = undefined,
 }: FunctionAPIContext<UpdateSessionAPIOptions<DefaultUser>>): Promise<UpdateSessionAPIReturn<DefaultUser>> => {
     try {
+        console.log("double submit token:", skipCSRFCheck || Boolean(doubleSubmitToken))
         const { session, headers } = await ctx.sessionStrategy.refreshSession(
             new Headers(headersInit),
             sessionInit,
-            skipCSRFCheck
+            skipCSRFCheck || Boolean(doubleSubmitToken)
         )
         if (!session) {
             throw new AuraAuthError({ code: "UPDATE_SESSION_INVALID" })
@@ -61,7 +63,8 @@ export const updateSession = async <DefaultUser extends User = User>({
             },
         } as UpdateSessionAPIReturn<DefaultUser>
     } catch (error) {
-        const { code, message } = handleApiError(error, "UPDATE_SESSION_INVALID", "Failed to update session.")
+        console.error("Error in updateSession API:", error)
+        const { code, message, statusCode } = handleApiError(error, "UPDATE_SESSION_INVALID", "Failed to update session.")
 
         const headers = new Headers(secureApiHeaders)
         return {
@@ -74,7 +77,7 @@ export const updateSession = async <DefaultUser extends User = User>({
             toResponse: () => {
                 return Response.json(
                     { success: false, session: null, redirect: false, redirectURL: null },
-                    { status: 400, headers }
+                    { status: statusCode, headers }
                 )
             },
         }