+++ packages/core/test/api/stateless/signOut.test.ts
@@ -118,4 +118,6 @@ describe("signOut API", async () => {
             toResponse: expect.any(Function),
         })
     })
+
+    test("")
 })